Move extracted modules from utils/ to services/ for coverage thresholds
server/utils/ requires 100% test coverage. These extracted modules
contain business logic (DB queries, HTTP calls) better classified
as services with an 80% threshold.
