+++ server/services/coverDownloader.js
